Все тесты прошли

+++ backend/recipes/serializers.py
@@ -62,7 +62,10 @@ class RecipeSerializer(serializers.ModelSerializer):
 
 
 class RecipeCreateSerializer(serializers.ModelSerializer):
-    ingredients = IngredientAmountSerializer(many=True)
+    ingredients = IngredientAmountSerializer(many=True, required=True)
+    tags = serializers.PrimaryKeyRelatedField(
+        queryset=Tag.objects.all(), many=True, required=True,
+    )
     image = Base64ImageField()
 
     class Meta:
@@ -72,6 +75,37 @@ class RecipeCreateSerializer(serializers.ModelSerializer):
             'name', 'image', 'text', 'cooking_time'
         )
         read_only_fields = ('id', 'author')
+
+    def validate_ingredients(self, value):
+        if not value:
+            raise serializers.ValidationError(
+                'Список ингредиентов не может быть пустым.'
+            )
+        ingredient_ids = [item['id'].id for item in value]
+        if len(ingredient_ids) != len(set(ingredient_ids)):
+            raise serializers.ValidationError(
+                'Ингредиенты не должны повторяться.'
+            )
+        return value
+
+    def validate_tags(self, value):
+        if not value:
+            raise serializers.ValidationError(
+                'Список тегов не может быть пустым.'
+            )
+        tag_ids = [tag.id for tag in value]
+        if len(tag_ids) != len(set(tag_ids)):
+            raise serializers.ValidationError(
+                'Теги не должны повторяться.'
+            )
+        return value
+
+    def validate_cooking_time(self, value):
+        if value < 1:
+            raise serializers.ValidationError(
+                'Время приготовления должно быть больше 0.'
+            )
+        return value
 
     def create(self, validated_data):
         ingredients_data = validated_data.pop('ingredients')
@@ -86,18 +120,32 @@ class RecipeCreateSerializer(serializers.ModelSerializer):
         return recipe
 
     def update(self, instance, validated_data):
-        ingredients_data = validated_data.pop('ingredients')
-        tags_data = validated_data.pop('tags')
+        ingredients_data = validated_data.pop('ingredients', None)
+        tags_data = validated_data.pop('tags', None)
+        if ingredients_data is None:
+            raise serializers.ValidationError(
+                {'ingredients': 'Обязательное поле.'}
+            )
+
+        if tags_data is None:
+            raise serializers.ValidationError(
+                {'tags': 'Обязательное поле.'}
+            )
+
         for attr, value in validated_data.items():
             setattr(instance, attr, value)
         instance.save()
-        instance.tags.set(tags_data)
-        instance.recipe_ingredients.all().delete()
-        RecipeIngredient.objects.bulk_create([
-            RecipeIngredient(recipe=instance, ingredient=i['id'],
-                             amount=i['amount'])
-            for i in ingredients_data
-        ])
+
+        if tags_data is not None:
+            instance.tags.set(tags_data)
+
+        if ingredients_data is not None:
+            instance.recipe_ingredients.all().delete()
+            RecipeIngredient.objects.bulk_create([
+                RecipeIngredient(recipe=instance, ingredient=i['id'],
+                                 amount=i['amount'])
+                for i in ingredients_data
+            ])
         return instance
 
     def to_representation(self, instance):
